refactor[login, register] separate logic of components
